+++ addon/components/bootstrap-inputs/-select.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+import layout from '../../templates/components/bootstrap-inputs/-select';
+
+export default Ember.Component.extend({
+  tagName: '',
+  layout
+});